Make iterm setup idempotent
[#155302614]

+++ setup-iterm2.sh
@@ -2,7 +2,6 @@
 current_directory=$PWD
 preferences_path="/Users/$(whoami)/Library/Preferences/com.googlecode.iterm2.plist"
 
-echo $current_directory
-echo $preferences_path
-
-ln -s "${current_directory}/assets/com.googlecode.iterm2.plist" "${preferences_path}"
+if ! [ -e "${preferences_path}" ]; then
+	ln -s "${current_directory}/assets/com.googlecode.iterm2.plist" "${preferences_path}"
+fi
